+++ www/Library/admin_widgets/RichTextEditor/Toolbar/Toolbar.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Панель инструментов для форматирования текста
+ *
+ * @version 1.0
+ * @date 11.04.2013
+ * @author Vladimir Shestakov <boolive@yandex.ru>
+ */
+namespace Library\admin_widgets\RichTextEditor\Toolbar;
+
+use Library\views\Widget\Widget,
+    Boolive\data\Data;
+
+class Toolbar extends Widget
+{
+    public function work($v = array())
+    {
+
+        return parent::work($v);
+    }
+}